backend/apisix_admin/requirements.txt
Django==3.0
django-auth-ldap==2.1.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.10.3
gunicorn==19.9.0
python-ldap==3.2.0
